slow PIT timer on MicrobeePP
Per EJW's notes about the original 2.4-based sources:

I also had to change the timer tick for the task switching to make it slower as the memory
subsystem is slow and so processing is also slow. As a result, the timer tick was going
off before it fully did a task switch without making the tick rate slower and it never got
to a command prompt.